+++ src/utils/InputHandler.js
@@ -1,0 +1,9 @@
+import { Console } from "@woowacourse/mission-utils";
+
+class InputHandler {
+    async getInput(message) {
+        return await Console.readLineAsync(message);
+    }
+}
+
+export default InputHandler;
